remove all configs except for genesis from testutil/config and consolidate into /config
