+++ tests/integration/precompiles/bech32/test_bech32.go
@@ -1,12 +1,12 @@
 package bech32
 
 import (
+	"github.com/cosmos/evm/config"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/vm"
 	"github.com/holiman/uint256"
 
 	"github.com/cosmos/evm/precompiles/bech32"
-	"github.com/cosmos/evm/testutil/config"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
