+++ testutil/ibc/helpers.go
@@ -1,6 +1,7 @@
 package ibctesting
 
 import (
+	"github.com/cosmos/evm/config"
 	"math/big"
 	"math/rand"
 	"testing"
@@ -9,8 +10,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	abci "github.com/cometbft/cometbft/abci/types"
-
-	"github.com/cosmos/evm/testutil/config"
 
 	bam "github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/client"
